+++ MelonLoader/Attributes/MelonInfoAttribute.cs
@@ -22,6 +22,11 @@ namespace MelonLoader
         public string Version { get; internal set; }
 
         /// <summary>
+        /// Semantic Version of the Melon. Will be null if Version is not using the <see href="https://semver.org">Semantic Versioning</see> format.
+        /// </summary>
+        public SemVersion SemanticVersion { get; internal set; }
+
+        /// <summary>
         /// Author of the Melon.
         /// </summary>
         public string Author { get; internal set; } // This used to be optional, but is now required
@@ -31,34 +36,87 @@ namespace MelonLoader
         /// </summary>
         public string DownloadLink { get; internal set; } // Might get Removed. Not sure yet.
 
-        internal MelonInfoAttribute(string name, string author, string version, string downloadLink, Type type)
+        /// <summary>
+        /// MelonInfo constructor.
+        /// </summary>
+        /// <param name="type">The main Melon type of the Melon (for example TestMod)</param>
+        /// <param name="name">Name of the Melon</param>
+        /// <param name="version">Version of the Melon</param>
+        /// <param name="author">Author of the Melon</param>
+        /// <param name="downloadLink">URL to the download link of the mod [optional]</param>
+        public MelonInfoAttribute(Type type, string name, string version, string author, string downloadLink = null) 
         {
-            SystemType = type;
-            Name = name ?? "UNKNOWN";
-            Version = version ?? "1.0.0";
-            Author = author ?? "UNKNOWN";
+            if (type == null)
+                throw new ArgumentNullException(nameof(type));
 
+            if (string.IsNullOrEmpty(name))
+                throw new ArgumentNullException(nameof(name));
+
+            if (string.IsNullOrEmpty(author))
+                throw new ArgumentNullException(nameof(author));
+
+            SystemType = type;
+            Name = name;
+            Author = author;
             DownloadLink = downloadLink; // Might get Removed. Not sure yet.
+
+            if (string.IsNullOrEmpty(version))
+                Version = "1.0.0";
+            else
+                Version = version;
+
+            if (SemVersion.TryParse(Version, out SemVersion semver))
+                SemanticVersion = semver;
         }
 
         /// <summary>
-        /// Main MelonInfo constructor.
+        /// MelonInfo constructor.
         /// </summary>
         /// <param name="type">The main Melon type of the Melon (for example TestMod)</param>
         /// <param name="name">Name of the Melon</param>
         /// <param name="version">Version of the Melon (Using the <see href="https://semver.org">Semantic Versioning</see> format)</param>
         /// <param name="author">Author of the Melon</param>
         /// <param name="downloadLink">URL to the download link of the mod [optional]</param>
-        public MelonInfoAttribute(Type type, string name, string version, string author, string downloadLink = null) 
+        public MelonInfoAttribute(Type type, string name, SemVersion version, string author, string downloadLink = null)
         {
+            if (type == null)
+                throw new ArgumentNullException(nameof(type));
+
+            if (string.IsNullOrEmpty(name))
+                throw new ArgumentNullException(nameof(name));
+
+            if (string.IsNullOrEmpty(author))
+                throw new ArgumentNullException(nameof(author));
+
+            if (version == null)
+                throw new ArgumentNullException(nameof(version));
+
             SystemType = type;
-            Name = string.IsNullOrEmpty(name) ? "UNKNOWN" : name;
-            Author = string.IsNullOrEmpty(author) ? "UNKNOWN" : author;
+            Name = name;
+            Version = version.ToString();
+            Author = author;
             DownloadLink = downloadLink; // Might get Removed. Not sure yet.
-
-            var isSemVer = SemVersion.TryParse(version, out SemVersion semver);
-
-            Version = isSemVer ? semver.ToString() : (string.IsNullOrEmpty(version) ? "1.0.0" : version);
         }
+        /// <summary>
+        /// MelonInfo constructor.
+        /// </summary>
+        /// <param name="type">The main Melon type of the Melon (for example TestMod)</param>
+        /// <param name="name">Name of the Melon</param>
+        /// <param name="version">Version of the Melon (Using the <see href="https://semver.org">Semantic Versioning</see> format)</param>
+        /// <param name="author">Author of the Melon</param>
+        /// <param name="downloadLink">URL to the download link of the mod [optional]</param>
+        public MelonInfoAttribute(Type type, string name, (int, int, int) version, string author, string downloadLink = null)
+            : this(type, name, (version.Item1, version.Item2, version.Item3, null), author, downloadLink) { }
+
+        /// <summary>
+        /// MelonInfo constructor.
+        /// </summary>
+        /// <param name="type">The main Melon type of the Melon (for example TestMod)</param>
+        /// <param name="name">Name of the Melon</param>
+        /// <param name="version">Version of the Melon (Using the <see href="https://semver.org">Semantic Versioning</see> format)</param>
+        /// <param name="author">Author of the Melon</param>
+        /// <param name="downloadLink">URL to the download link of the mod [optional]</param>
+        public MelonInfoAttribute(Type type, string name, (int, int, int, string) version, string author, string downloadLink = null)
+            : this(type, name, $"{version.Item1}.{version.Item2}.{version.Item3}{(string.IsNullOrEmpty(version.Item4) ? "" : version.Item4)}", author, downloadLink) { }
     }
 }